refactor: split Vaillant semantic runtime files
