+++ cmd/gateway/semantic_vaillant.go
@@ -403,1766 +403,6 @@ var b524CapabilityProbes = []b524ProbeSpec{
 	{opcode: localDHW.opcode, group: localDHW.group, instance: dhwInstance, addr: dhw_operation_mode},
 }
 
-type zonePresenceState string
-
-const (
-	zonePresenceAbsent           zonePresenceState = "ABSENT"
-	zonePresenceSuspectResurrect zonePresenceState = "SUSPECT_RESURRECT"
-	zonePresencePresent          zonePresenceState = "PRESENT"
-	zonePresenceSuspectMissing   zonePresenceState = "SUSPECT_MISSING"
-)
-
-type zonePresenceRecord struct {
-	State      zonePresenceState
-	HitStreak  int
-	MissStreak int
-}
-
-type vaillantZoneSnapshot struct {
-	Instance byte
-	Present  bool
-
-	Name string
-
-	OperatingMode string
-	Preset        string
-	HvacAction    string
-	AllowedModes  []string
-
-	CurrentTempC *float64
-	TargetTempC  *float64
-	HumidityPct  *float64
-
-	ConfigurationHeatingOperationMode          string
-	StateSpecialFunction                       string
-	ConfigurationRoomTemperatureZoneMappingRaw *uint16
-	ConfigurationAssociatedCircuitRaw          *uint16
-	ConfigurationCircuitTypeRaw                *uint16
-	StateValveStatusRaw                        *uint16
-
-	QuickVetoTempC     *float64
-	QuickVetoDurationH *float64
-	QuickVetoEndTime   string // "HH:MM" or "" if no veto
-	QuickVetoEndDate   string // "YYYY-MM-DD" or "" if no veto
-
-	HolidayStartDate string   // "YYYY-MM-DD" or ""
-	HolidayEndDate   string   // "YYYY-MM-DD" or ""
-	HolidaySetpointC *float64 // configured setpoint during holiday
-	HolidayStartTime string   // "HH:MM" or ""
-	HolidayEndTime   string   // "HH:MM" or ""
-
-	FieldFreshness map[semanticFieldKey]semanticFieldFreshness
-}
-
-type vaillantDhwSnapshot struct {
-	OperatingMode string
-	Preset        string
-
-	CurrentTempC *float64
-	TargetTempC  *float64
-
-	ConfigurationDHWOperationMode string
-	StateSpecialFunction          string
-
-	HolidayStartDate string // "YYYY-MM-DD" or ""
-	HolidayEndDate   string // "YYYY-MM-DD" or ""
-
-	FieldFreshness map[semanticFieldKey]semanticFieldFreshness
-}
-
-type vaillantCircuitSnapshot struct {
-	Instance   byte
-	Active     bool
-	Controller byte
-
-	CircuitTypeRaw     *uint16
-	CoolingEnabledRaw  *uint16
-	FlowSetpointC      *float64
-	FlowTemperatureC   *float64
-	HeatingCurve       *float64
-	FlowTempMaxC       *float64
-	FlowTempMinC       *float64
-	SummerLimitC       *float64
-	RoomTempControlRaw *uint16
-	CircuitStateRaw    *uint16
-	CircuitStateLiveAt time.Time
-	FrostProtectionC   *float64
-	PumpStatusRaw      *uint16
-	PumpStatusLiveAt   time.Time
-	CalcFlowTempC      *float64
-	MixerPositionPct   *float64
-	HumidityPct        *float64
-	DewPointC          *float64
-	PumpHoursRaw       *uint32
-	PumpStartsRaw      *uint32
-}
-
-type radioDeviceKey struct {
-	Group    byte
-	Instance byte
-}
-
-type vaillantRadioDeviceSnapshot struct {
-	Group                byte
-	Instance             byte
-	SlotMode             string
-	DeviceConnected      *bool
-	DeviceClassAddress   *uint8
-	DeviceModel          string
-	FirmwareVersion      *string
-	HardwareIdentifier   *uint16
-	RemoteControlAddress *uint8
-	DevicePaired         *bool
-	ReceptionStrength    *uint8
-	ZoneAssignment       *uint8
-	RoomTemperatureC     *float64
-	RoomHumidityPct      *float64
-}
-
-type vaillantSolarSnapshot struct {
-	CollectorTemperatureC *float64
-	ReturnTemperatureC    *float64
-	PumpActive            *bool
-	CurrentYield          *float64
-	PumpHours             *uint32
-	SolarEnabled          *bool
-	FunctionMode          *bool
-}
-
-type vaillantCylinderSnapshot struct {
-	Instance         byte
-	TemperatureC     *float64
-	MaxSetpointC     *float64
-	ChargeHysteresis *float64
-	ChargeOffset     *float64
-}
-
-type semanticSnapshotSource uint8
-
-const (
-	semanticSnapshotSourceCache semanticSnapshotSource = iota
-	semanticSnapshotSourceLive
-)
-
-type semanticFieldKey string
-
-const (
-	zoneFieldName                          semanticFieldKey = "zone.name"
-	zoneFieldOperatingMode                 semanticFieldKey = "zone.operating_mode"
-	zoneFieldPreset                        semanticFieldKey = "zone.preset"
-	zoneFieldHvacAction                    semanticFieldKey = "zone.hvac_action"
-	zoneFieldAllowedModes                  semanticFieldKey = "zone.allowed_modes"
-	zoneFieldCurrentTempC                  semanticFieldKey = "zone.current_temp_c"
-	zoneFieldTargetTempC                   semanticFieldKey = "zone.target_temp_c"
-	zoneFieldCurrentHumidityPct            semanticFieldKey = "zone.current_humidity_pct"
-	zoneFieldSpecialFunctionRaw            semanticFieldKey = "zone.special_function_raw"
-	zoneFieldZoneOperationModeRaw          semanticFieldKey = "zone.operation_mode_raw"
-	zoneFieldRoomTemperatureZoneMappingRaw semanticFieldKey = "zone.room_temperature_zone_mapping_raw"
-	zoneFieldZoneCircuitIndexRaw           semanticFieldKey = "zone.circuit_index_raw"
-	zoneFieldCircuitTypeRaw                semanticFieldKey = "zone.circuit_type_raw"
-	zoneFieldZoneValveStatusRaw            semanticFieldKey = "zone.valve_status_raw"
-	zoneFieldQuickVetoTempC                semanticFieldKey = "zone.quick_veto_temp_c"
-	zoneFieldQuickVetoDurationH            semanticFieldKey = "zone.quick_veto_duration_h"
-	zoneFieldQuickVetoEndTime              semanticFieldKey = "zone.quick_veto_end_time"
-	zoneFieldQuickVetoEndDate              semanticFieldKey = "zone.quick_veto_end_date"
-	zoneFieldHolidayStartDate              semanticFieldKey = "zone.holiday_start_date"
-	zoneFieldHolidayEndDate                semanticFieldKey = "zone.holiday_end_date"
-	zoneFieldHolidaySetpointC              semanticFieldKey = "zone.holiday_setpoint_c"
-	zoneFieldHolidayStartTime              semanticFieldKey = "zone.holiday_start_time"
-	zoneFieldHolidayEndTime                semanticFieldKey = "zone.holiday_end_time"
-	dhwFieldOperatingMode                  semanticFieldKey = "dhw.operating_mode"
-	dhwFieldPreset                         semanticFieldKey = "dhw.preset"
-	dhwFieldCurrentTempC                   semanticFieldKey = "dhw.current_temp_c"
-	dhwFieldTargetTempC                    semanticFieldKey = "dhw.target_temp_c"
-	dhwFieldSpecialFunctionRaw             semanticFieldKey = "dhw.special_function_raw"
-	dhwFieldDhwOperationModeRaw            semanticFieldKey = "dhw.operation_mode_raw"
-	dhwFieldHolidayStartDate               semanticFieldKey = "dhw.holiday_start_date"
-	dhwFieldHolidayEndDate                 semanticFieldKey = "dhw.holiday_end_date"
-)
-
-type semanticFieldFreshness struct {
-	Source semanticSnapshotSource
-	Stale  bool
-}
-
-type semanticFieldSet map[semanticFieldKey]struct{}
-
-func newSemanticFieldSet(keys ...semanticFieldKey) semanticFieldSet {
-	set := make(semanticFieldSet, len(keys))
-	for _, key := range keys {
-		set[key] = struct{}{}
-	}
-	return set
-}
-
-func (set semanticFieldSet) has(key semanticFieldKey) bool {
-	if len(set) == 0 {
-		return false
-	}
-	_, ok := set[key]
-	return ok
-}
-
-type boilerStatusTier uint8
-
-const (
-	boilerStatusTierFast boilerStatusTier = iota
-	boilerStatusTierMedium
-	boilerStatusTierSlow
-)
-
-type boilerStatusTierSchedule struct {
-	tier     boilerStatusTier
-	interval time.Duration
-	priority semanticTaskPriority
-}
-
-const (
-	semanticTaskRefreshRegulatorCapability semanticTaskKey = "refresh_regulator_capability"
-	semanticTaskRefreshDiscovery           semanticTaskKey = "refresh_discovery"
-	semanticTaskRefreshConfig              semanticTaskKey = "refresh_config"
-	semanticTaskRefreshState               semanticTaskKey = "refresh_state"
-	semanticTaskRefreshCircuits            semanticTaskKey = "refresh_circuits"
-	semanticTaskRefreshSystem              semanticTaskKey = "refresh_system"
-	semanticTaskRefreshRadioDevices        semanticTaskKey = "refresh_radio_devices"
-	semanticTaskRefreshEnergy              semanticTaskKey = "refresh_energy"
-	semanticTaskRefreshSchedules           semanticTaskKey = "refresh_schedules"
-	semanticTaskRefreshBoilerFast          semanticTaskKey = "refresh_boiler_fast"
-	semanticTaskRefreshBoilerMedium        semanticTaskKey = "refresh_boiler_medium"
-	semanticTaskRefreshBoilerSlow          semanticTaskKey = "refresh_boiler_slow"
-)
-
-var (
-	zoneConfigRefreshFieldSet = newSemanticFieldSet(
-		zoneFieldName,
-		zoneFieldOperatingMode,
-		zoneFieldPreset,
-		zoneFieldAllowedModes,
-		zoneFieldTargetTempC,
-		zoneFieldCurrentHumidityPct,
-		zoneFieldZoneOperationModeRaw,
-		zoneFieldRoomTemperatureZoneMappingRaw,
-		zoneFieldZoneCircuitIndexRaw,
-		zoneFieldCircuitTypeRaw,
-		zoneFieldQuickVetoTempC,
-		zoneFieldQuickVetoDurationH,
-		zoneFieldQuickVetoEndTime,
-		zoneFieldQuickVetoEndDate,
-		zoneFieldHolidayStartDate,
-		zoneFieldHolidayEndDate,
-		zoneFieldHolidaySetpointC,
-		zoneFieldHolidayStartTime,
-		zoneFieldHolidayEndTime,
-	)
-	zoneFastStateFieldSet = newSemanticFieldSet(
-		zoneFieldOperatingMode,
-		zoneFieldPreset,
-		zoneFieldHvacAction,
-		zoneFieldAllowedModes,
-		zoneFieldCurrentTempC,
-		zoneFieldSpecialFunctionRaw,
-		zoneFieldZoneValveStatusRaw,
-	)
-	zoneGrabFieldSet = newSemanticFieldSet(
-		zoneFieldName,
-		zoneFieldOperatingMode,
-		zoneFieldPreset,
-		zoneFieldHvacAction,
-		zoneFieldAllowedModes,
-		zoneFieldCurrentTempC,
-		zoneFieldTargetTempC,
-		zoneFieldCurrentHumidityPct,
-		zoneFieldZoneOperationModeRaw,
-		zoneFieldSpecialFunctionRaw,
-		zoneFieldRoomTemperatureZoneMappingRaw,
-		zoneFieldZoneCircuitIndexRaw,
-		zoneFieldZoneValveStatusRaw,
-		zoneFieldCircuitTypeRaw,
-	)
-	dhwFieldSet = newSemanticFieldSet(
-		dhwFieldOperatingMode,
-		dhwFieldPreset,
-		dhwFieldCurrentTempC,
-		dhwFieldTargetTempC,
-		dhwFieldDhwOperationModeRaw,
-		dhwFieldSpecialFunctionRaw,
-		dhwFieldHolidayStartDate,
-		dhwFieldHolidayEndDate,
-	)
-)
-
-func newVaillantSemanticPoller(cfg ebusgateway.Config, gateway *ebusgateway.Gateway, provider *graphql.LiveSemanticProvider, hub *graphql.BroadcastHub, cache semanticCachePersister) *vaillantSemanticPoller {
-	catalog, catalogErr := productids.LoadCatalog()
-	observeFirstFlags := normalizeObserveFirstFeatureFlagsForPoller(cfg)
-	shadow := ebusgateway.NewShadowCache(ebusgateway.ShadowCacheOptions{
-		FeatureFlags: observeFirstFlags,
-	})
-	poller := &vaillantSemanticPoller{
-		scheduler:       ebusgateway.NewSemanticReadScheduler(),
-		tasks:           newSemanticTaskScheduler(),
-		reg:             gateway.Registry,
-		bus:             gateway.Bus,
-		provider:        provider,
-		hub:             hub,
-		cache:           cache,
-		shadow:          shadow,
-		transportConfig: cfg.TransportConfig,
-		watchObserver:   cfg.WatchObserver,
-		watchEfficiency: cfg.WatchEfficiencyObserver,
-		source:          cfg.ScanSource,
-		companion:       cfg.StartupCompanionTarget,
-
-		requestTimeout:           cfg.SemanticRequestTimeout,
-		discoveryInterval:        cfg.SemanticDiscoveryInterval,
-		configInterval:           cfg.SemanticConfigInterval,
-		stateInterval:            cfg.SemanticStateInterval,
-		energyInterval:           cfg.SemanticEnergyInterval,
-		scheduleInterval:         10 * time.Minute,
-		boilerFastInterval:       30 * time.Second,
-		boilerMediumInterval:     5 * time.Minute,
-		boilerSlowInterval:       10 * time.Minute,
-		zoneMissThreshold:        cfg.SemanticZonePresenceMissThreshold,
-		zoneHitThreshold:         cfg.SemanticZonePresenceHitThreshold,
-		dhwStaleTTL:              cfg.SemanticDHWStaleTTL,
-		deviceSlotRediscoveryTTL: 30 * time.Minute,
-		fm5EvidenceTTL:           defaultFM5EvidenceTTL,
-		circuitFullScanInterval:  semanticCircuitFullScanInterval,
-
-		catalog:    catalog,
-		catalogErr: catalogErr,
-
-		regulatorRecheckInterval: cfg.SemanticRegulatorRecheckInterval,
-		regulatorAbsenceGrace:    cfg.SemanticRegulatorAbsenceGrace,
-		regAbsenceState:          regulatorPresent,
-		fm5Mode:                  graphql.Fm5SemanticModeAbsent,
-
-		zones:          make(map[byte]*vaillantZoneSnapshot),
-		presence:       make(map[byte]*zonePresenceRecord),
-		circuits:       make(map[byte]*vaillantCircuitSnapshot),
-		radioDevices:   make(map[radioDeviceKey]*vaillantRadioDeviceSnapshot),
-		solarCylinders: make(map[byte]*vaillantCylinderSnapshot),
-		nowFn:          time.Now,
-	}
-	if poller.zoneMissThreshold <= 0 {
-		poller.zoneMissThreshold = ebusgateway.DefaultSemanticZonePresenceMissThreshold
-	}
-	if poller.zoneHitThreshold <= 0 {
-		poller.zoneHitThreshold = ebusgateway.DefaultSemanticZonePresenceHitThreshold
-	}
-	if poller.dhwStaleTTL <= 0 {
-		poller.dhwStaleTTL = ebusgateway.DefaultSemanticDHWStaleTTL
-	}
-	if poller.energyInterval <= 0 {
-		poller.energyInterval = ebusgateway.DefaultSemanticEnergyInterval
-	}
-	if poller.regulatorRecheckInterval <= 0 {
-		poller.regulatorRecheckInterval = ebusgateway.DefaultSemanticRegulatorRecheckInterval
-	}
-	if poller.regulatorAbsenceGrace <= 0 {
-		poller.regulatorAbsenceGrace = ebusgateway.DefaultSemanticRegulatorAbsenceGrace
-	}
-	semanticZoneCount.Set(0)
-	semanticRegulatorState.Set(string(poller.regAbsenceState))
-	poller.scheduler.SetShadowCache(shadow)
-	poller.scheduler.SetCircuitBreaker(ebusgateway.SemanticReadCircuitBreakerOptions{
-		FailureBudget:      cfg.SemanticReadBreakerFailureBudget,
-		OpenCooldown:       cfg.SemanticReadBreakerOpenCooldown,
-		HalfOpenProbeLimit: cfg.SemanticReadBreakerHalfOpenProbeLimit,
-		OnTransition:       poller.onSemanticReadBreakerTransition,
-		OnSuppressed:       poller.onSemanticReadBreakerSuppressed,
-	})
-	poller.adapterInfo = newVaillantAdapterInfoState(gateway.Bus, gateway.Transport, provider)
-	poller.routerPlanesRefreshFn = func() { _ = gateway.RefreshRouterPlanes() }
-	return poller
-}
-
-func normalizeObserveFirstFeatureFlagsForPoller(cfg ebusgateway.Config) ebusgateway.ObserveFirstFeatureFlags {
-	if cfg.ExternalWritePolicy == "" &&
-		!cfg.ObserveFirstEnabled &&
-		!cfg.PassiveStateDirectApply &&
-		!cfg.PassiveConfigDirectApply {
-		return ebusgateway.NormalizeObserveFirstFeatureFlagsFromView(cfg.ObserveFirstFlags)
-	}
-	return ebusgateway.NormalizeObserveFirstFeatureFlags(
-		cfg.ObserveFirstEnabled,
-		cfg.PassiveStateDirectApply,
-		cfg.PassiveConfigDirectApply,
-		cfg.ExternalWritePolicy,
-	)
-}
-
-func (p *vaillantSemanticPoller) onSemanticReadBreakerTransition(event ebusgateway.SemanticReadCircuitBreakerTransition) {
-	if p == nil {
-		return
-	}
-	semanticReadBreakerTransitionsTotal.Add(fmt.Sprintf("%s->%s", event.From, event.To), 1)
-	log.Printf(
-		"semantic_read_breaker_transition key=%q from=%s to=%s consecutive_failures=%d",
-		event.Key,
-		event.From,
-		event.To,
-		event.ConsecutiveFailures,
-	)
-}
-
-func (p *vaillantSemanticPoller) onSemanticReadBreakerSuppressed(event ebusgateway.SemanticReadCircuitBreakerSuppression) {
-	if p == nil {
-		return
-	}
-	semanticReadBreakerSuppressedTotal.Add(string(event.State), 1)
-	log.Printf(
-		"semantic_read_breaker_suppressed key=%q state=%s retry_after=%s suppressed_total=%d",
-		event.Key,
-		event.State,
-		event.RetryAfter.Round(time.Millisecond),
-		event.SuppressedTotal,
-	)
-}
-
-func preloadSemanticCache(provider *graphql.LiveSemanticProvider, cacheStore *semanticCacheStore) (semanticCacheSnapshot, bool) {
-	if provider == nil || cacheStore == nil {
-		return semanticCacheSnapshot{}, false
-	}
-	snapshot, ok := cacheStore.Load()
-	if !ok {
-		return semanticCacheSnapshot{}, false
-	}
-	if len(snapshot.Zones) > 0 {
-		provider.SetZonesFromCache(snapshot.Zones)
-	}
-	if snapshot.DHW != nil {
-		provider.SetDHWFromCache(snapshot.DHW)
-	}
-	if snapshot.Boiler != nil {
-		provider.SetBoilerStatusFromCache(snapshot.Boiler)
-	}
-	return snapshot, true
-}
-
-func (p *vaillantSemanticPoller) hydrateFromCache(snapshot semanticCacheSnapshot) {
-	if p == nil {
-		return
-	}
-
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	if p.zones == nil {
-		p.zones = make(map[byte]*vaillantZoneSnapshot)
-	}
-	if p.presence == nil {
-		p.presence = make(map[byte]*zonePresenceRecord)
-	}
-	for idx, zone := range snapshot.Zones {
-		instance := zoneInstanceFromSemanticID(zone.ID, idx)
-		p.zones[instance] = zoneSnapshotFromSemanticZone(instance, zone)
-		p.presence[instance] = &zonePresenceRecord{
-			State:     zonePresencePresent,
-			HitStreak: p.zoneHitThresholdValue(),
-		}
-	}
-	if snapshot.DHW != nil {
-		p.dhw = dhwSnapshotFromSemanticStatus(snapshot.DHW)
-		if !snapshot.PersistedAt.IsZero() {
-			p.dhwLastUpdateAt = snapshot.PersistedAt.UTC()
-		} else {
-			p.markDHWUpdatedNowLocked()
-		}
-	}
-	if snapshot.Boiler != nil {
-		p.boiler = boilerSnapshotFromGraphQL(snapshot.Boiler)
-	}
-	semanticZoneCount.Set(int64(len(p.zones)))
-}
-
-func zoneInstanceFromSemanticID(id string, fallbackIndex int) byte {
-	trimmed := strings.TrimSpace(strings.ToLower(id))
-	if strings.HasPrefix(trimmed, "zone-") {
-		ordinalRaw := strings.TrimPrefix(trimmed, "zone-")
-		if ordinal, err := strconv.Atoi(ordinalRaw); err == nil && ordinal > 0 && ordinal <= 256 {
-			return byte(ordinal - 1)
-		}
-	}
-	if fallbackIndex < 0 {
-		return 0
-	}
-	if fallbackIndex > 255 {
-		return 255
-	}
-	return byte(fallbackIndex)
-}
-
-func zoneSnapshotFromSemanticZone(instance byte, zone graphql.Zone) *vaillantZoneSnapshot {
-	snapshot := &vaillantZoneSnapshot{
-		Instance:             instance,
-		Present:              true,
-		Name:                 zone.Name,
-		OperatingMode:        zone.Config.OperatingMode,
-		Preset:               zone.Config.Preset,
-		HvacAction:           zone.State.HvacAction,
-		AllowedModes:         append([]string(nil), zone.Config.AllowedModes...),
-		CurrentTempC:         cloneFloat64Ptr(zone.State.CurrentTempC),
-		TargetTempC:          cloneFloat64Ptr(zone.Config.TargetTempC),
-		HumidityPct:          cloneFloat64Ptr(zone.State.CurrentHumidityPct),
-		StateSpecialFunction: zone.State.SpecialFunction,
-	}
-	if zone.Config.AssociatedCircuit != nil {
-		v := uint16(*zone.Config.AssociatedCircuit)
-		snapshot.ConfigurationAssociatedCircuitRaw = &v
-	}
-	if zone.Config.RoomTemperatureZoneMapping != nil {
-		v := uint16(*zone.Config.RoomTemperatureZoneMapping)
-		snapshot.ConfigurationRoomTemperatureZoneMappingRaw = &v
-	}
-	if zone.Config.CircuitType != "" {
-		v := encodeCircuitTypeRaw(zone.Config.CircuitType)
-		snapshot.ConfigurationCircuitTypeRaw = &v
-	}
-	if zone.State.ValvePositionPct != nil {
-		v := uint16(*zone.State.ValvePositionPct * 655.35)
-		snapshot.StateValveStatusRaw = &v
-	}
-	snapshot.QuickVetoTempC = cloneFloat64Ptr(zone.Config.QuickVetoSetpointC)
-	snapshot.QuickVetoDurationH = cloneFloat64Ptr(zone.Config.QuickVetoDurationH)
-	snapshot.HolidayStartDate = zone.Config.HolidayStartDate
-	snapshot.HolidayEndDate = zone.Config.HolidayEndDate
-	snapshot.HolidaySetpointC = cloneFloat64Ptr(zone.Config.HolidaySetpointC)
-	snapshot.HolidayStartTime = zone.Config.HolidayStartTime
-	snapshot.HolidayEndTime = zone.Config.HolidayEndTime
-	seedZoneFreshness(snapshot, semanticSnapshotSourceCache, true)
-	return snapshot
-}
-
-func encodeCircuitTypeRaw(circuitType string) uint16 {
-	switch circuitType {
-	case "radiator":
-		return 0
-	case "underfloor":
-		return 1
-	case "mixed":
-		return 2
-	default:
-		var n uint16
-		if _, err := fmt.Sscanf(circuitType, "unknown_%d", &n); err == nil {
-			return n
-		}
-		return 0
-	}
-}
-
-func dhwSnapshotFromSemanticStatus(status *graphql.DhwStatus) *vaillantDhwSnapshot {
-	if status == nil {
-		return nil
-	}
-	snapshot := &vaillantDhwSnapshot{
-		OperatingMode:        status.Config.OperatingMode,
-		Preset:               status.Config.Preset,
-		CurrentTempC:         cloneFloat64Ptr(status.State.CurrentTempC),
-		TargetTempC:          cloneFloat64Ptr(status.Config.TargetTempC),
-		StateSpecialFunction: status.State.SpecialFunction,
-		HolidayStartDate:     status.Config.HolidayStartDate,
-		HolidayEndDate:       status.Config.HolidayEndDate,
-	}
-	seedDhwFreshness(snapshot, semanticSnapshotSourceCache, true)
-	return snapshot
-}
-
-func cloneFloat64Ptr(value *float64) *float64 {
-	if value == nil {
-		return nil
-	}
-	copyValue := *value
-	return &copyValue
-}
-
-func cloneUint16Ptr(value *uint16) *uint16 {
-	if value == nil {
-		return nil
-	}
-	copyValue := *value
-	return &copyValue
-}
-
-func cloneUint8Ptr(value *uint8) *uint8 {
-	if value == nil {
-		return nil
-	}
-	copyValue := *value
-	return &copyValue
-}
-
-func cloneUint32Ptr(value *uint32) *uint32 {
-	if value == nil {
-		return nil
-	}
-	copyValue := *value
-	return &copyValue
-}
-
-func ensureFieldFreshnessMap(entry *map[semanticFieldKey]semanticFieldFreshness) map[semanticFieldKey]semanticFieldFreshness {
-	if *entry == nil {
-		*entry = make(map[semanticFieldKey]semanticFieldFreshness)
-	}
-	return *entry
-}
-
-func setFieldFreshness(fields map[semanticFieldKey]semanticFieldFreshness, field semanticFieldKey, source semanticSnapshotSource, stale bool) {
-	if fields == nil {
-		return
-	}
-	fields[field] = semanticFieldFreshness{
-		Source: source,
-		Stale:  stale,
-	}
-}
-
-func markFieldStale(fields map[semanticFieldKey]semanticFieldFreshness, field semanticFieldKey) {
-	if fields == nil {
-		return
-	}
-	freshness, ok := fields[field]
-	if !ok {
-		return
-	}
-	freshness.Stale = true
-	fields[field] = freshness
-}
-
-func mergeStringField(target *string, incoming string, attempted bool, fields map[semanticFieldKey]semanticFieldFreshness, field semanticFieldKey, source semanticSnapshotSource) {
-	if !attempted {
-		return
-	}
-	if strings.TrimSpace(incoming) != "" {
-		*target = incoming
-		setFieldFreshness(fields, field, source, false)
-		return
-	}
-	markFieldStale(fields, field)
-}
-
-func mergeStringSliceField(target *[]string, incoming []string, attempted bool, fields map[semanticFieldKey]semanticFieldFreshness, field semanticFieldKey, source semanticSnapshotSource) {
-	if !attempted {
-		return
-	}
-	if len(incoming) > 0 {
-		*target = append([]string(nil), incoming...)
-		setFieldFreshness(fields, field, source, false)
-		return
-	}
-	markFieldStale(fields, field)
-}
-
-func mergeFloatField(target **float64, incoming *float64, attempted bool, fields map[semanticFieldKey]semanticFieldFreshness, field semanticFieldKey, source semanticSnapshotSource) {
-	if !attempted {
-		return
-	}
-	if incoming != nil {
-		*target = cloneFloat64Ptr(incoming)
-		setFieldFreshness(fields, field, source, false)
-		return
-	}
-	markFieldStale(fields, field)
-}
-
-func mergeUint16Field(target **uint16, incoming *uint16, attempted bool, fields map[semanticFieldKey]semanticFieldFreshness, field semanticFieldKey, source semanticSnapshotSource) {
-	if !attempted {
-		return
-	}
-	if incoming != nil {
-		*target = cloneUint16Ptr(incoming)
-		setFieldFreshness(fields, field, source, false)
-		return
-	}
-	markFieldStale(fields, field)
-}
-
-func mergeZoneSnapshotFields(entry *vaillantZoneSnapshot, incoming *vaillantZoneSnapshot, source semanticSnapshotSource, attempted semanticFieldSet) {
-	if entry == nil || incoming == nil {
-		return
-	}
-	fields := ensureFieldFreshnessMap(&entry.FieldFreshness)
-
-	mergeStringField(&entry.Name, incoming.Name, attempted.has(zoneFieldName), fields, zoneFieldName, source)
-	mergeStringField(&entry.OperatingMode, incoming.OperatingMode, attempted.has(zoneFieldOperatingMode), fields, zoneFieldOperatingMode, source)
-	mergeStringField(&entry.Preset, incoming.Preset, attempted.has(zoneFieldPreset), fields, zoneFieldPreset, source)
-	mergeStringField(&entry.HvacAction, incoming.HvacAction, attempted.has(zoneFieldHvacAction), fields, zoneFieldHvacAction, source)
-	mergeStringSliceField(&entry.AllowedModes, incoming.AllowedModes, attempted.has(zoneFieldAllowedModes), fields, zoneFieldAllowedModes, source)
-	mergeFloatField(&entry.CurrentTempC, incoming.CurrentTempC, attempted.has(zoneFieldCurrentTempC), fields, zoneFieldCurrentTempC, source)
-	mergeFloatField(&entry.TargetTempC, incoming.TargetTempC, attempted.has(zoneFieldTargetTempC), fields, zoneFieldTargetTempC, source)
-	mergeFloatField(&entry.HumidityPct, incoming.HumidityPct, attempted.has(zoneFieldCurrentHumidityPct), fields, zoneFieldCurrentHumidityPct, source)
-	mergeStringField(&entry.ConfigurationHeatingOperationMode, incoming.ConfigurationHeatingOperationMode, attempted.has(zoneFieldZoneOperationModeRaw), fields, zoneFieldZoneOperationModeRaw, source)
-	mergeStringField(&entry.StateSpecialFunction, incoming.StateSpecialFunction, attempted.has(zoneFieldSpecialFunctionRaw), fields, zoneFieldSpecialFunctionRaw, source)
-	mergeUint16Field(&entry.ConfigurationRoomTemperatureZoneMappingRaw, incoming.ConfigurationRoomTemperatureZoneMappingRaw, attempted.has(zoneFieldRoomTemperatureZoneMappingRaw), fields, zoneFieldRoomTemperatureZoneMappingRaw, source)
-	mergeUint16Field(&entry.ConfigurationAssociatedCircuitRaw, incoming.ConfigurationAssociatedCircuitRaw, attempted.has(zoneFieldZoneCircuitIndexRaw), fields, zoneFieldZoneCircuitIndexRaw, source)
-	mergeUint16Field(&entry.ConfigurationCircuitTypeRaw, incoming.ConfigurationCircuitTypeRaw, attempted.has(zoneFieldCircuitTypeRaw), fields, zoneFieldCircuitTypeRaw, source)
-	mergeUint16Field(&entry.StateValveStatusRaw, incoming.StateValveStatusRaw, attempted.has(zoneFieldZoneValveStatusRaw), fields, zoneFieldZoneValveStatusRaw, source)
-	mergeFloatField(&entry.QuickVetoTempC, incoming.QuickVetoTempC, attempted.has(zoneFieldQuickVetoTempC), fields, zoneFieldQuickVetoTempC, source)
-	mergeFloatField(&entry.QuickVetoDurationH, incoming.QuickVetoDurationH, attempted.has(zoneFieldQuickVetoDurationH), fields, zoneFieldQuickVetoDurationH, source)
-	mergeStringField(&entry.QuickVetoEndTime, incoming.QuickVetoEndTime, attempted.has(zoneFieldQuickVetoEndTime), fields, zoneFieldQuickVetoEndTime, source)
-	mergeStringField(&entry.QuickVetoEndDate, incoming.QuickVetoEndDate, attempted.has(zoneFieldQuickVetoEndDate), fields, zoneFieldQuickVetoEndDate, source)
-	mergeStringField(&entry.HolidayStartDate, incoming.HolidayStartDate, attempted.has(zoneFieldHolidayStartDate), fields, zoneFieldHolidayStartDate, source)
-	mergeStringField(&entry.HolidayEndDate, incoming.HolidayEndDate, attempted.has(zoneFieldHolidayEndDate), fields, zoneFieldHolidayEndDate, source)
-	mergeFloatField(&entry.HolidaySetpointC, incoming.HolidaySetpointC, attempted.has(zoneFieldHolidaySetpointC), fields, zoneFieldHolidaySetpointC, source)
-	mergeStringField(&entry.HolidayStartTime, incoming.HolidayStartTime, attempted.has(zoneFieldHolidayStartTime), fields, zoneFieldHolidayStartTime, source)
-	mergeStringField(&entry.HolidayEndTime, incoming.HolidayEndTime, attempted.has(zoneFieldHolidayEndTime), fields, zoneFieldHolidayEndTime, source)
-}
-
-func mergeDhwSnapshotFields(entry *vaillantDhwSnapshot, incoming *vaillantDhwSnapshot, source semanticSnapshotSource, attempted semanticFieldSet) {
-	if entry == nil || incoming == nil {
-		return
-	}
-	fields := ensureFieldFreshnessMap(&entry.FieldFreshness)
-
-	mergeStringField(&entry.OperatingMode, incoming.OperatingMode, attempted.has(dhwFieldOperatingMode), fields, dhwFieldOperatingMode, source)
-	mergeStringField(&entry.Preset, incoming.Preset, attempted.has(dhwFieldPreset), fields, dhwFieldPreset, source)
-	mergeFloatField(&entry.CurrentTempC, incoming.CurrentTempC, attempted.has(dhwFieldCurrentTempC), fields, dhwFieldCurrentTempC, source)
-	mergeFloatField(&entry.TargetTempC, incoming.TargetTempC, attempted.has(dhwFieldTargetTempC), fields, dhwFieldTargetTempC, source)
-	mergeStringField(&entry.ConfigurationDHWOperationMode, incoming.ConfigurationDHWOperationMode, attempted.has(dhwFieldDhwOperationModeRaw), fields, dhwFieldDhwOperationModeRaw, source)
-	mergeStringField(&entry.StateSpecialFunction, incoming.StateSpecialFunction, attempted.has(dhwFieldSpecialFunctionRaw), fields, dhwFieldSpecialFunctionRaw, source)
-	mergeStringField(&entry.HolidayStartDate, incoming.HolidayStartDate, attempted.has(dhwFieldHolidayStartDate), fields, dhwFieldHolidayStartDate, source)
-	mergeStringField(&entry.HolidayEndDate, incoming.HolidayEndDate, attempted.has(dhwFieldHolidayEndDate), fields, dhwFieldHolidayEndDate, source)
-}
-
-func seedZoneFreshness(snapshot *vaillantZoneSnapshot, source semanticSnapshotSource, stale bool) {
-	if snapshot == nil {
-		return
-	}
-	fields := ensureFieldFreshnessMap(&snapshot.FieldFreshness)
-	if strings.TrimSpace(snapshot.Name) != "" {
-		setFieldFreshness(fields, zoneFieldName, source, stale)
-	}
-	if strings.TrimSpace(snapshot.OperatingMode) != "" {
-		setFieldFreshness(fields, zoneFieldOperatingMode, source, stale)
-	}
-	if strings.TrimSpace(snapshot.Preset) != "" {
-		setFieldFreshness(fields, zoneFieldPreset, source, stale)
-	}
-	if strings.TrimSpace(snapshot.HvacAction) != "" {
-		setFieldFreshness(fields, zoneFieldHvacAction, source, stale)
-	}
-	if len(snapshot.AllowedModes) > 0 {
-		setFieldFreshness(fields, zoneFieldAllowedModes, source, stale)
-	}
-	if snapshot.CurrentTempC != nil {
-		setFieldFreshness(fields, zoneFieldCurrentTempC, source, stale)
-	}
-	if snapshot.TargetTempC != nil {
-		setFieldFreshness(fields, zoneFieldTargetTempC, source, stale)
-	}
-	if snapshot.HumidityPct != nil {
-		setFieldFreshness(fields, zoneFieldCurrentHumidityPct, source, stale)
-	}
-	if strings.TrimSpace(snapshot.ConfigurationHeatingOperationMode) != "" {
-		setFieldFreshness(fields, zoneFieldZoneOperationModeRaw, source, stale)
-	}
-	if strings.TrimSpace(snapshot.StateSpecialFunction) != "" {
-		setFieldFreshness(fields, zoneFieldSpecialFunctionRaw, source, stale)
-	}
-	if snapshot.ConfigurationRoomTemperatureZoneMappingRaw != nil {
-		setFieldFreshness(fields, zoneFieldRoomTemperatureZoneMappingRaw, source, stale)
-	}
-	if snapshot.ConfigurationAssociatedCircuitRaw != nil {
-		setFieldFreshness(fields, zoneFieldZoneCircuitIndexRaw, source, stale)
-	}
-	if snapshot.ConfigurationCircuitTypeRaw != nil {
-		setFieldFreshness(fields, zoneFieldCircuitTypeRaw, source, stale)
-	}
-	if snapshot.StateValveStatusRaw != nil {
-		setFieldFreshness(fields, zoneFieldZoneValveStatusRaw, source, stale)
-	}
-	if snapshot.QuickVetoTempC != nil {
-		setFieldFreshness(fields, zoneFieldQuickVetoTempC, source, stale)
-	}
-	if snapshot.QuickVetoDurationH != nil {
-		setFieldFreshness(fields, zoneFieldQuickVetoDurationH, source, stale)
-	}
-	if snapshot.QuickVetoEndTime != "" {
-		setFieldFreshness(fields, zoneFieldQuickVetoEndTime, source, stale)
-	}
-	if snapshot.QuickVetoEndDate != "" {
-		setFieldFreshness(fields, zoneFieldQuickVetoEndDate, source, stale)
-	}
-	if snapshot.HolidayStartDate != "" {
-		setFieldFreshness(fields, zoneFieldHolidayStartDate, source, stale)
-	}
-	if snapshot.HolidayEndDate != "" {
-		setFieldFreshness(fields, zoneFieldHolidayEndDate, source, stale)
-	}
-	if snapshot.HolidaySetpointC != nil {
-		setFieldFreshness(fields, zoneFieldHolidaySetpointC, source, stale)
-	}
-	if snapshot.HolidayStartTime != "" {
-		setFieldFreshness(fields, zoneFieldHolidayStartTime, source, stale)
-	}
-	if snapshot.HolidayEndTime != "" {
-		setFieldFreshness(fields, zoneFieldHolidayEndTime, source, stale)
-	}
-}
-
-func seedDhwFreshness(snapshot *vaillantDhwSnapshot, source semanticSnapshotSource, stale bool) {
-	if snapshot == nil {
-		return
-	}
-	fields := ensureFieldFreshnessMap(&snapshot.FieldFreshness)
-	if strings.TrimSpace(snapshot.OperatingMode) != "" {
-		setFieldFreshness(fields, dhwFieldOperatingMode, source, stale)
-	}
-	if strings.TrimSpace(snapshot.Preset) != "" {
-		setFieldFreshness(fields, dhwFieldPreset, source, stale)
-	}
-	if snapshot.CurrentTempC != nil {
-		setFieldFreshness(fields, dhwFieldCurrentTempC, source, stale)
-	}
-	if snapshot.TargetTempC != nil {
-		setFieldFreshness(fields, dhwFieldTargetTempC, source, stale)
-	}
-	if strings.TrimSpace(snapshot.ConfigurationDHWOperationMode) != "" {
-		setFieldFreshness(fields, dhwFieldDhwOperationModeRaw, source, stale)
-	}
-	if strings.TrimSpace(snapshot.StateSpecialFunction) != "" {
-		setFieldFreshness(fields, dhwFieldSpecialFunctionRaw, source, stale)
-	}
-	if snapshot.HolidayStartDate != "" {
-		setFieldFreshness(fields, dhwFieldHolidayStartDate, source, stale)
-	}
-	if snapshot.HolidayEndDate != "" {
-		setFieldFreshness(fields, dhwFieldHolidayEndDate, source, stale)
-	}
-}
-
-func (p *vaillantSemanticPoller) boilerStatusTierSchedules() []boilerStatusTierSchedule {
-	if p == nil {
-		return nil
-	}
-	return []boilerStatusTierSchedule{
-		{tier: boilerStatusTierFast, interval: p.boilerFastInterval, priority: semanticTaskPriorityHigh},
-		{tier: boilerStatusTierMedium, interval: p.boilerMediumInterval, priority: semanticTaskPriorityMedium},
-		{tier: boilerStatusTierSlow, interval: p.boilerSlowInterval, priority: semanticTaskPriorityLow},
-	}
-}
-
-func (p *vaillantSemanticPoller) boilerStatusTierTask(tier boilerStatusTier) func(context.Context) {
-	return func(ctx context.Context) {
-		if tier == boilerStatusTierFast {
-			p.refreshBoilerStatus(ctx)
-			return
-		}
-		p.refreshBoilerStatusTier(ctx, tier)
-	}
-}
-
-func boilerStatusTierTaskKey(tier boilerStatusTier) semanticTaskKey {
-	switch tier {
-	case boilerStatusTierFast:
-		return semanticTaskRefreshBoilerFast
-	case boilerStatusTierMedium:
-		return semanticTaskRefreshBoilerMedium
-	case boilerStatusTierSlow:
-		return semanticTaskRefreshBoilerSlow
-	default:
-		return semanticTaskKey("")
-	}
-}
-
-func (p *vaillantSemanticPoller) enqueueBoilerStatusPriming(ctx context.Context) {
-	if p == nil {
-		return
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if p.tasks == nil {
-		for _, schedule := range p.boilerStatusTierSchedules() {
-			p.boilerStatusTierTask(schedule.tier)(ctx)
-		}
-		return
-	}
-	for _, schedule := range p.boilerStatusTierSchedules() {
-		p.enqueueTask(boilerStatusTierTaskKey(schedule.tier), schedule.priority, p.boilerStatusTierTask(schedule.tier))
-	}
-}
-
-func (p *vaillantSemanticPoller) enqueueControllerSemanticPriming(ctx context.Context) {
-	if p == nil {
-		return
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if p.tasks == nil {
-		p.refreshConfig(ctx)
-		p.refreshCircuits(ctx)
-		p.refreshSystem(ctx)
-		p.refreshRadioDevices(ctx)
-		p.refreshEnergy(ctx)
-		return
-	}
-	p.enqueueTask(semanticTaskRefreshConfig, semanticTaskPriorityHigh, p.refreshConfig)
-	p.enqueueTask(semanticTaskRefreshCircuits, semanticTaskPriorityMedium, p.refreshCircuits)
-	p.enqueueTask(semanticTaskRefreshSystem, semanticTaskPriorityMedium, p.refreshSystem)
-	p.enqueueTask(semanticTaskRefreshRadioDevices, semanticTaskPriorityMedium, p.refreshRadioDevices)
-	p.enqueueTask(semanticTaskRefreshEnergy, semanticTaskPriorityMedium, p.refreshEnergy)
-}
-
-func (p *vaillantSemanticPoller) refreshStartupSemanticPlanes(ctx context.Context) {
-	if p == nil {
-		return
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-
-	p.mu.Lock()
-	controller := p.controller
-	p.mu.Unlock()
-	if controller == 0 {
-		p.refreshDiscovery(ctx)
-		p.mu.Lock()
-		controller = p.controller
-		p.mu.Unlock()
-	}
-	if controller == 0 {
-		return
-	}
-	if p.provider == nil {
-		return
-	}
-
-	p.publishStartupSchedules()
-	p.publishStartupPlaneDefaults()
-	primingCtx := ctx
-	cancel := func() {}
-	if semanticStartupPrimingBudget > 0 {
-		primingCtx, cancel = context.WithTimeout(ctx, semanticStartupPrimingBudget)
-	}
-	defer cancel()
-
-	attempts := 0
-	for {
-		attempts++
-		status := p.startupL1PrimingStatus()
-		if !status.dhw {
-			p.refreshDHWStartupUntilReady(primingCtx, 1)
-			status = p.startupL1PrimingStatus()
-		}
-		if !status.zones {
-			p.refreshZoneDiscovery(primingCtx, true)
-			status = p.startupL1PrimingStatus()
-		}
-		if !status.dhw {
-			p.refreshDHWStartupUntilReady(primingCtx, 1)
-		}
-		if !status.system {
-			p.refreshSystemStartup(primingCtx)
-		}
-		if !status.boilerStatus {
-			p.refreshBoilerStatusStartup(primingCtx)
-		}
-		status = p.startupL1PrimingStatus()
-		if !status.circuits {
-			p.refreshCircuitsStartup(primingCtx)
-		}
-		if !status.radioDevices {
-			p.refreshRadioDevicesStartup(primingCtx)
-		}
-		status = p.startupL1PrimingStatus()
-		if !status.fm5Satisfied && status.fm5Evidence && !status.fm5GateKnown {
-			p.refreshSystemStartup(primingCtx)
-			status = p.startupL1PrimingStatus()
-		}
-		if status.system && (!status.fm5Satisfied || !status.solar || !status.cylinders) {
-			p.refreshFM5SemanticStartup(primingCtx)
-		}
-		status = p.startupL1PrimingStatus()
-		if status.ready() {
-			log.Printf("semantic_startup_l1_priming result=ready attempts=%d %s", attempts, status.String())
-			return
-		}
-		if semanticStartupPrimingBudget <= 0 {
-			log.Printf("semantic_startup_l1_priming result=incomplete attempts=%d %s", attempts, status.String())
-			return
-		}
-
-		delay := semanticStartupPrimingRetryDelay
-		if delay <= 0 {
-			delay = 50 * time.Millisecond
-		}
-		timer := time.NewTimer(delay)
-		select {
-		case <-primingCtx.Done():
-			if !timer.Stop() {
-				select {
-				case <-timer.C:
-				default:
-				}
-			}
-			log.Printf("semantic_startup_l1_priming result=deadline attempts=%d %s", attempts, status.String())
-			return
-		case <-timer.C:
-		}
-	}
-}
-
-func (p *vaillantSemanticPoller) publishStartupPlaneDefaults() {
-	if p == nil || p.provider == nil {
-		return
-	}
-	if p.provider.Circuits() == nil {
-		p.provider.SetCircuits([]graphql.CircuitStatus{})
-	}
-	if p.provider.Solar() == nil {
-		p.provider.SetSolar(&graphql.SolarStatus{})
-	}
-	if p.provider.Cylinders() == nil {
-		p.provider.SetCylinders([]graphql.CylinderStatus{})
-	}
-}
-
-func (p *vaillantSemanticPoller) refreshStartupCriticalSemanticPlanes(ctx context.Context) {
-	if p == nil || p.provider == nil {
-		return
-	}
-
-	p.publishStartupSchedules()
-	status := p.startupL1PrimingStatus()
-	if !status.dhw {
-		p.refreshDHWStartupUntilReady(ctx, semanticStartupCriticalDHWAttempts)
-	}
-	status = p.startupL1PrimingStatus()
-	if !status.system {
-		p.refreshSystemStartup(ctx)
-	}
-	status = p.startupL1PrimingStatus()
-	if !status.boilerStatus {
-		p.refreshBoilerStatusStartup(ctx)
-	}
-}
-
-func (p *vaillantSemanticPoller) refreshDHWStartupUntilReady(ctx context.Context, attempts int) bool {
-	if p == nil {
-		return false
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if attempts <= 0 {
-		attempts = 1
-	}
-	for attempt := 0; attempt < attempts; attempt++ {
-		if p.startupL1PrimingStatus().dhw {
-			return true
-		}
-		p.refreshDHWStartup(ctx)
-		if p.startupL1PrimingStatus().dhw {
-			return true
-		}
-		if attempt == attempts-1 {
-			break
-		}
-		delay := semanticStartupPrimingRetryDelay
-		if delay <= 0 {
-			delay = 50 * time.Millisecond
-		}
-		timer := time.NewTimer(delay)
-		select {
-		case <-ctx.Done():
-			if !timer.Stop() {
-				select {
-				case <-timer.C:
-				default:
-				}
-			}
-			return false
-		case <-timer.C:
-		}
-	}
-	return p.startupL1PrimingStatus().dhw
-}
-
-type startupL1PrimingStatus struct {
-	zones        bool
-	dhw          bool
-	circuits     bool
-	system       bool
-	radioDevices bool
-	fm5GateKnown bool
-	fm5Evidence  bool
-	fm5Required  bool
-	fm5Satisfied bool
-	solar        bool
-	cylinders    bool
-	boilerStatus bool
-}
-
-func (status startupL1PrimingStatus) ready() bool {
-	return status.zones &&
-		status.dhw &&
-		status.circuits &&
-		status.system &&
-		status.radioDevices &&
-		status.fm5Satisfied &&
-		status.boilerStatus
-}
-
-func (status startupL1PrimingStatus) String() string {
-	return fmt.Sprintf(
-		"zones=%t dhw=%t circuits=%t system=%t radio_devices=%t fm5_gate_known=%t fm5_evidence=%t fm5_required=%t fm5_satisfied=%t solar=%t cylinders=%t boiler_status=%t",
-		status.zones,
-		status.dhw,
-		status.circuits,
-		status.system,
-		status.radioDevices,
-		status.fm5GateKnown,
-		status.fm5Evidence,
-		status.fm5Required,
-		status.fm5Satisfied,
-		status.solar,
-		status.cylinders,
-		status.boilerStatus,
-	)
-}
-
-func (p *vaillantSemanticPoller) startupL1PrimingStatus() startupL1PrimingStatus {
-	if p == nil || p.provider == nil {
-		return startupL1PrimingStatus{}
-	}
-	p.mu.Lock()
-	moduleConfig := (*uint16)(nil)
-	if p.system != nil {
-		moduleConfig = cloneUint16Ptr(p.system.ModuleConfigurationVR71)
-	}
-	fm5GateKnown := p.system != nil && p.system.ModuleConfigurationVR71 != nil
-	radioSnapshots := make([]*vaillantRadioDeviceSnapshot, 0, len(p.radioDevices))
-	for _, snapshot := range p.radioDevices {
-		if snapshot != nil {
-			radioSnapshots = append(radioSnapshots, cloneRadioSnapshot(snapshot))
-		}
-	}
-	radioProbed := p.startupRadioDevicesProbed
-	fm5RegistryEvidenceIgnored := p.fm5RegistryEvidenceIgnored
-	p.mu.Unlock()
-	liveFM5Evidence := hasFM5EvidenceFromRadioSnapshots(radioSnapshots)
-	fm5Evidence := liveFM5Evidence || (!fm5RegistryEvidenceIgnored && p.hasFM5RegistryEvidence())
-	fm5Required := fm5Evidence && moduleConfig != nil && *moduleConfig <= 2
-	zonesReady := p.provider.Zones() != nil
-	solarReady := p.provider.Solar() != nil
-	cylinders := p.provider.Cylinders()
-	cylindersPublished := cylinders != nil
-	interpretedCylindersReady := len(cylinders) > 0
-	fm5Mode := p.provider.FM5SemanticMode()
-	fm5NonInterpretedPublished := fm5Evidence &&
-		fm5Mode == graphql.Fm5SemanticModeGPIOOnly &&
-		solarReady &&
-		cylindersPublished
-	fm5Satisfied := !fm5Evidence ||
-		fm5NonInterpretedPublished ||
-		(fm5GateKnown && !fm5Required) ||
-		(fm5Required && solarReady && interpretedCylindersReady)
-	return startupL1PrimingStatus{
-		zones:        zonesReady,
-		dhw:          p.provider.DHW() != nil,
-		circuits:     p.provider.Circuits() != nil,
-		system:       p.provider.System() != nil,
-		radioDevices: radioProbed || len(p.provider.RadioDevices()) > 0,
-		fm5GateKnown: fm5GateKnown,
-		fm5Evidence:  fm5Evidence,
-		fm5Required:  fm5Required,
-		fm5Satisfied: fm5Satisfied,
-		solar:        solarReady,
-		cylinders:    cylindersPublished,
-		boilerStatus: p.provider.BoilerStatus() != nil,
-	}
-}
-
-func (p *vaillantSemanticPoller) startupProbeContext(ctx context.Context) (context.Context, context.CancelFunc) {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	return context.WithTimeout(ctx, semanticStartupProbeTimeout)
-}
-
-func (p *vaillantSemanticPoller) readB524Startup(ctx context.Context, opcode, group, instance byte, addr uint16) ([]byte, bool) {
-	probeCtx, cancel := p.startupProbeContext(ctx)
-	defer cancel()
-	return p.readB524ValueLive(probeCtx, opcode, group, instance, addr)
-}
-
-func (p *vaillantSemanticPoller) readB524Uint16Startup(ctx context.Context, opcode, group, instance byte, addr uint16) (*uint16, bool) {
-	raw, ok := p.readB524Startup(ctx, opcode, group, instance, addr)
-	if !ok {
-		return nil, false
-	}
-	value, ok := decodeB524Uint16(raw)
-	if !ok {
-		return nil, false
-	}
-	return &value, true
-}
-
-func (p *vaillantSemanticPoller) publishStartupSchedules() {
-	if p == nil || p.provider == nil || p.provider.Schedules() != nil {
-		return
-	}
-	p.provider.SetSchedules(&graphql.ScheduleStatus{Programs: []graphql.ScheduleProgram{}})
-}
-
-func (p *vaillantSemanticPoller) refreshDHWStartup(ctx context.Context) {
-	if p == nil {
-		return
-	}
-
-	attempted := make(semanticFieldSet)
-	status := &vaillantDhwSnapshot{}
-	readAny := false
-	if raw, ok := p.readB524Startup(ctx, localDHW.opcode, localDHW.group, dhwInstance, dhw_current_temp); ok {
-		if value := decodeB524Float32FromRaw(raw); value != nil {
-			status.CurrentTempC = value
-			attempted[dhwFieldCurrentTempC] = struct{}{}
-			readAny = true
-		}
-	}
-	opModeRaw, opModeOK := p.readB524Uint16Startup(ctx, localDHW.opcode, localDHW.group, dhwInstance, dhw_operation_mode)
-	if opModeOK {
-		attempted[dhwFieldOperatingMode] = struct{}{}
-		attempted[dhwFieldPreset] = struct{}{}
-		attempted[dhwFieldDhwOperationModeRaw] = struct{}{}
-		readAny = true
-	}
-	if opModeOK {
-		status.OperatingMode, status.Preset = deriveDhwModeAndPreset(opModeRaw, nil)
-		status.ConfigurationDHWOperationMode = formatUintToken(*opModeRaw)
-	}
-
-	p.mu.Lock()
-	source := semanticSnapshotSourceCache
-	if readAny {
-		if p.dhw == nil {
-			p.dhw = &vaillantDhwSnapshot{}
-		}
-		mergeDhwSnapshotFields(p.dhw, status, semanticSnapshotSourceLive, attempted)
-		p.markDHWUpdatedNowLocked()
-		source = semanticSnapshotSourceLive
-	}
-	hasSnapshot := p.dhw != nil
-	p.mu.Unlock()
-	if hasSnapshot {
-		p.publishDHW(source)
-	}
-}
-
-func (p *vaillantSemanticPoller) refreshCircuitsStartup(ctx context.Context) {
-	if p == nil {
-		return
-	}
-
-	p.mu.Lock()
-	controller := p.controller
-	p.mu.Unlock()
-	instances := allCircuitRefreshInstances()
-	updates := make(map[byte]*vaillantCircuitSnapshot)
-	inactive := make(map[byte]struct{})
-	for _, instance := range instances {
-		circuitTypeRaw, ok := p.readB524Uint16Startup(ctx, localCircuits.opcode, localCircuits.group, instance, circuit_type)
-		if !ok || circuitTypeRaw == nil {
-			continue
-		}
-		switch *circuitTypeRaw {
-		case 0x0000, 0x00FF, 0xFFFF:
-			if snapshot, known := p.readDhwPseudoCircuitStartupEvidence(ctx, instance); snapshot != nil {
-				snapshot.Controller = controller
-				updates[instance] = snapshot
-				continue
-			} else if !known && instance == dhwPseudoCircuitInstance {
-				continue
-			}
-			inactive[instance] = struct{}{}
-			continue
-		default:
-			updates[instance] = &vaillantCircuitSnapshot{
-				Instance:       instance,
-				Active:         true,
-				Controller:     controller,
-				CircuitTypeRaw: cloneUint16Ptr(circuitTypeRaw),
-			}
-		}
-	}
-	if len(updates) == 0 && len(inactive) == 0 {
-		p.mu.Lock()
-		hasExisting := len(p.circuits) > 0
-		p.mu.Unlock()
-		if hasExisting {
-			p.publishCircuits(semanticSnapshotSourceCache)
-		}
-		return
-	}
-
-	p.mu.Lock()
-	if p.circuits == nil {
-		p.circuits = make(map[byte]*vaillantCircuitSnapshot)
-	}
-	for instance := range inactive {
-		delete(p.circuits, instance)
-	}
-	for instance, incoming := range updates {
-		if incoming.CircuitStateRaw != nil {
-			incoming.CircuitStateLiveAt = p.now()
-		}
-		if incoming.PumpStatusRaw != nil {
-			incoming.PumpStatusLiveAt = p.now()
-		}
-		p.circuits[instance] = mergeCircuitSnapshotNonDestructive(p.circuits[instance], incoming)
-	}
-	p.lastCircuitFullScanAt = p.now()
-	p.lastCircuitFullScanComplete = len(updates)+len(inactive) == len(instances)
-	p.mu.Unlock()
-	p.publishCircuits(semanticSnapshotSourceLive)
-}
-
-func (p *vaillantSemanticPoller) readDhwPseudoCircuitStartupEvidence(ctx context.Context, instance byte) (*vaillantCircuitSnapshot, bool) {
-	if p == nil || instance != dhwPseudoCircuitInstance {
-		return nil, false
-	}
-	snapshot := newDhwPseudoCircuitSnapshot(instance)
-	flow, flowKnown := p.readDhwPseudoCircuitStartupTemperatureEvidence(ctx, instance, circuit_flow_temp)
-	if isDhwPseudoCircuitTemperatureEvidence(flow) {
-		snapshot.FlowTemperatureC = flow
-	}
-	calc, calcKnown := p.readDhwPseudoCircuitStartupTemperatureEvidence(ctx, instance, circuit_calc_flow_temp)
-	if isDhwPseudoCircuitTemperatureEvidence(calc) {
-		snapshot.CalcFlowTempC = calc
-	}
-	if snapshot.FlowTemperatureC == nil && snapshot.CalcFlowTempC == nil {
-		return nil, flowKnown && calcKnown
-	}
-	return snapshot, true
-}
-
-func (p *vaillantSemanticPoller) readDhwPseudoCircuitStartupTemperatureEvidence(ctx context.Context, instance byte, addr uint16) (*float64, bool) {
-	raw, ok := p.readB524Startup(ctx, localCircuits.opcode, localCircuits.group, instance, addr)
-	if !ok {
-		return nil, false
-	}
-	return decodeB524Float32FromRaw(raw), true
-}
-
-func (p *vaillantSemanticPoller) refreshSystemStartup(ctx context.Context) {
-	if p == nil {
-		return
-	}
-
-	p.mu.Lock()
-	controller := p.controller
-	p.mu.Unlock()
-	snapshot := &vaillantSystemSnapshot{Controller: controller}
-	readAny := false
-	configurationAcquired := false
-	if raw, ok := p.readB524Uint16Startup(ctx, localRegulator.opcode, localRegulator.group, regulatorInstance, system_scheme); ok && raw != nil {
-		snapshot.SystemScheme = cloneUint16Ptr(raw)
-		readAny = true
-	}
-	if raw, ok := p.readB524Uint16Startup(ctx, localRegulator.opcode, localRegulator.group, regulatorInstance, system_module_configuration_vr71); ok && raw != nil {
-		snapshot.ModuleConfigurationVR71 = cloneUint16Ptr(raw)
-		readAny = true
-		configurationAcquired = true
-	}
-	if raw, ok := p.readB524Startup(ctx, localRegulator.opcode, localRegulator.group, regulatorInstance, system_water_pressure); ok {
-		if value := decodeB524Float32FromRaw(raw); value != nil {
-			snapshot.SystemWaterPressure = value
-			readAny = true
-		}
-	}
-
-	p.mu.Lock()
-	source := semanticSnapshotSourceCache
-	p.updateFM5ConfigurationAcquisitionLocked(configurationAcquired)
-	if readAny {
-		if snapshot.SystemFlowTemperature != nil {
-			snapshot.SystemFlowTemperatureLiveAt = p.now()
-		}
-		p.updateSystemSnapshotLocked(mergeSystemSnapshotNonDestructive(p.system, snapshot))
-		source = semanticSnapshotSourceLive
-	}
-	hasSnapshot := p.system != nil
-	p.mu.Unlock()
-	if hasSnapshot {
-		p.publishSystem(source)
-	}
-}
-
-func (p *vaillantSemanticPoller) refreshFM5ConfigurationForLateIdentity(ctx context.Context) {
-	if p == nil {
-		return
-	}
-
-	p.mu.Lock()
-	controller := p.controller
-	p.mu.Unlock()
-	snapshot := &vaillantSystemSnapshot{Controller: controller}
-	configurationAcquired := false
-	if raw, ok := p.readB524Uint16Startup(ctx, localRegulator.opcode, localRegulator.group, regulatorInstance, system_module_configuration_vr71); ok && raw != nil {
-		snapshot.ModuleConfigurationVR71 = cloneUint16Ptr(raw)
-		configurationAcquired = true
-	}
-
-	p.mu.Lock()
-	source := semanticSnapshotSourceCache
-	p.updateFM5ConfigurationAcquisitionLocked(configurationAcquired)
-	if configurationAcquired {
-		p.updateSystemSnapshotLocked(mergeSystemSnapshotNonDestructive(p.system, snapshot))
-		source = semanticSnapshotSourceLive
-	}
-	hasSnapshot := p.system != nil
-	p.mu.Unlock()
-	if hasSnapshot {
-		p.publishSystem(source)
-	}
-}
-
-func (p *vaillantSemanticPoller) refreshRadioDevicesStartup(ctx context.Context) {
-	if p == nil {
-		return
-	}
-
-	discovered := p.registryRadioDeviceSeeds()
-	fullScanGroups := startupRadioFullScanGroups(discovered)
-	fm5PositiveObserved := false
-	verified := make(map[radioDeviceKey]bool)
-	readAny := false
-	probeSlot := func(grp b524GroupDef, instance byte) {
-		connectedRaw, ok := p.readB524Startup(ctx, grp.opcode, grp.group, instance, device_slot_connected)
-		if !ok || len(connectedRaw) == 0 {
-			return
-		}
-		readAny = true
-		connected := connectedRaw[0] == 1
-		classAddress := p.readB524U8Startup(ctx, grp.opcode, grp.group, instance, device_slot_class_address)
-		key := radioDeviceKey{Group: grp.group, Instance: instance}
-		verified[key] = true
-		include, slotMode := startupRadioDeviceInclude(grp.group, connected, classAddress)
-		if !include {
-			delete(discovered, key)
-			return
-		}
-		discovered[key] = &vaillantRadioDeviceSnapshot{
-			Group:              grp.group,
-			Instance:           instance,
-			SlotMode:           slotMode,
-			DeviceConnected:    &connected,
-			DeviceClassAddress: cloneUint8Ptr(classAddress),
-			DeviceModel:        decodeRadioDeviceModel(classAddress),
-		}
-	}
-
-	for _, grp := range remoteDeviceGroups {
-		if grp.group == remoteFunctionalModules.group {
-			continue
-		}
-		for instance := byte(0x00); instance <= semanticStartupSlotFastMaxInstance; instance++ {
-			probeSlot(grp, instance)
-		}
-	}
-	fm5LastInstance := semanticStartupSlotFastMaxInstance
-	if fullScanGroups[remoteFunctionalModules.group] {
-		fm5LastInstance = semanticStartupSlotFullMaxInstance
-	}
-	fm5Scan := p.scanFunctionalModuleIdentityBudgetAwareWith(ctx, 0x00, fm5LastInstance, p.readB524Startup)
-	if fm5Scan.observedAny {
-		readAny = true
-	}
-	for instance := byte(0x00); instance <= fm5LastInstance; instance++ {
-		key := radioDeviceKey{Group: remoteFunctionalModules.group, Instance: instance}
-		verified[key] = true
-		if snapshot := fm5Scan.snapshots[instance]; snapshot != nil {
-			discovered[key] = snapshot
-		} else {
-			delete(discovered, key)
-		}
-	}
-	fm5PositiveObserved = fm5Scan.positiveSnapshot != nil
-	fm5NamespaceComplete := fullScanGroups[remoteFunctionalModules.group] && fm5Scan.namespaceComplete
-	if !fm5PositiveObserved {
-		for _, grp := range remoteDeviceGroups {
-			if grp.group == remoteFunctionalModules.group || !fullScanGroups[grp.group] {
-				continue
-			}
-			for instance := semanticStartupSlotFastMaxInstance + 1; instance <= semanticStartupSlotFullMaxInstance; instance++ {
-				probeSlot(grp, instance)
-			}
-		}
-	}
-	if readAny {
-		for key, snapshot := range discovered {
-			if snapshot != nil && snapshot.SlotMode == "registry" && !verified[key] {
-				if fm5PositiveObserved && key.Group != remoteFunctionalModules.group {
-					continue
-				}
-				delete(discovered, key)
-			}
-		}
-	}
-
-	liveFM5Evidence := hasFM5EvidenceFromRadioMap(discovered)
-	p.mu.Lock()
-	p.startupRadioDevicesProbed = readAny
-	p.fm5IdentityScanComplete = fm5NamespaceComplete
-	p.fm5IdentityIncoherent = !fm5NamespaceComplete && !liveFM5Evidence
-	if fm5NamespaceComplete {
-		p.fm5RegistryEvidenceIgnored = !liveFM5Evidence
-	} else if liveFM5Evidence {
-		p.fm5RegistryEvidenceIgnored = false
-	}
-	if len(discovered) == 0 {
-		if readAny {
-			p.radioDevices = make(map[radioDeviceKey]*vaillantRadioDeviceSnapshot)
-			p.fm5EvidenceGeneration++
-			if fm5NamespaceComplete {
-				p.fm5IdentityObservedAt = p.now()
-			}
-		}
-		p.mu.Unlock()
-		if readAny {
-			p.publishRadioDevices(semanticSnapshotSourceLive)
-		}
-		return
-	}
-	if p.radioDevices == nil || readAny {
-		p.radioDevices = make(map[radioDeviceKey]*vaillantRadioDeviceSnapshot)
-	}
-	for key, snapshot := range discovered {
-		p.radioDevices[key] = snapshot
-	}
-	if readAny {
-		p.fm5EvidenceGeneration++
-		if fm5NamespaceComplete || liveFM5Evidence {
-			p.fm5IdentityObservedAt = p.now()
-		}
-	}
-	p.mu.Unlock()
-	source := semanticSnapshotSourceCache
-	if readAny {
-		source = semanticSnapshotSourceLive
-	}
-	p.publishRadioDevices(source)
-}
-
-func startupRadioFullScanGroups(discovered map[radioDeviceKey]*vaillantRadioDeviceSnapshot) map[byte]bool {
-	seeded := make(map[byte]bool)
-	highSeeded := make(map[byte]bool)
-	for key := range discovered {
-		seeded[key.Group] = true
-		if key.Instance > semanticStartupSlotFastMaxInstance {
-			highSeeded[key.Group] = true
-		}
-	}
-
-	out := make(map[byte]bool)
-	for _, grp := range remoteDeviceGroups {
-		if !seeded[grp.group] || highSeeded[grp.group] {
-			out[grp.group] = true
-		}
-	}
-	return out
-}
-
-func startupRadioDeviceInclude(group byte, connected bool, classAddress *uint8) (bool, string) {
-	switch group {
-	case remoteRegulators.group, remoteThermostats.group:
-		return connected, "startup"
-	case remoteFunctionalModules.group:
-		if connected {
-			return true, "startup"
-		}
-		if classAddress != nil && *classAddress == circuitManagingDeviceVR71Address {
-			return true, "inventory"
-		}
-		return false, "startup"
-	default:
-		return connected, "startup"
-	}
-}
-
-func (p *vaillantSemanticPoller) registryRadioDeviceSeeds() map[radioDeviceKey]*vaillantRadioDeviceSnapshot {
-	out := make(map[radioDeviceKey]*vaillantRadioDeviceSnapshot)
-	if p == nil || p.reg == nil {
-		return out
-	}
-	nextInstance := map[byte]byte{}
-	p.reg.IterateSnapshots(func(snap registry.DeviceEntrySnapshot) bool {
-		addr := ebusgateway.SnapshotTargetAddressForRouting(snap)
-		deviceID := normalizeDeviceID(snap.DeviceID)
-		var group byte
-		switch {
-		case strings.HasPrefix(deviceID, "BASV") || addr == 0x15:
-			group = remoteRegulators.group
-		case strings.HasPrefix(deviceID, "VR71") || strings.HasPrefix(deviceID, "FM5") || addr == circuitManagingDeviceVR71Address:
-			group = remoteFunctionalModules.group
-		case strings.HasPrefix(deviceID, "VR92"):
-			group = remoteThermostats.group
-		default:
-			return true
-		}
-		instance := nextInstance[group]
-		nextInstance[group] = instance + 1
-		classAddress := uint8(addr)
-		connected := true
-		out[radioDeviceKey{Group: group, Instance: instance}] = &vaillantRadioDeviceSnapshot{
-			Group:              group,
-			Instance:           instance,
-			SlotMode:           "registry",
-			DeviceConnected:    &connected,
-			DeviceClassAddress: &classAddress,
-			DeviceModel:        decodeRadioDeviceModel(&classAddress),
-		}
-		return true
-	})
-	return out
-}
-
-func (p *vaillantSemanticPoller) readB524U8Startup(ctx context.Context, opcode, group, instance byte, addr uint16) *uint8 {
-	raw, ok := p.readB524Startup(ctx, opcode, group, instance, addr)
-	if !ok || len(raw) == 0 {
-		return nil
-	}
-	value := raw[0]
-	return &value
-}
-
-func (p *vaillantSemanticPoller) refreshFM5SemanticStartup(ctx context.Context) {
-	if p == nil || p.provider == nil {
-		return
-	}
-
-	evidence := p.captureFM5Evidence()
-	moduleConfig := evidence.moduleConfig
-	if evidence.configurationFailed {
-		moduleConfig = nil
-	}
-	fm5GateSatisfied := moduleConfig != nil && *moduleConfig <= 2
-	observedAt := p.now()
-	evidenceStale := evidence.staleAt(observedAt, p.fm5EvidenceTTL)
-	var incomingSolar *vaillantSolarSnapshot
-	incomingCylinders := make(map[byte]*vaillantCylinderSnapshot)
-	solarReadable := false
-	cylindersReadable := false
-	if evidence.controller != 0 && fm5GateSatisfied && !evidenceStale && !evidence.identityIncoherent {
-		incomingSolar, solarReadable = p.readSolarSnapshotStartup(ctx)
-		incomingCylinders, cylindersReadable = p.readCylinderSnapshotsStartup(ctx)
-	}
-	currentEvidence := p.captureFM5Evidence()
-	incoherent := evidence.identityIncoherent || currentEvidence.identityIncoherent || !evidence.sameGeneration(currentEvidence)
-	evidenceRevision := p.nextFM5EvidenceRevision(evidence.generation, currentEvidence.generation)
-	negativeIdentityObserved := evidence.hasNegativeObservation() && currentEvidence.hasNegativeObservation()
-	freshNegativeIdentityObserved := evidence.hasFreshNegativeObservation(observedAt, p.fm5EvidenceTTL) &&
-		currentEvidence.hasFreshNegativeObservation(observedAt, p.fm5EvidenceTTL)
-	var verdict graphql.Fm5Interpretation
-	if freshNegativeIdentityObserved && !incoherent && evidence.controller != 0 && moduleConfig != nil {
-		verdict = graphql.Fm5Interpretation{
-			Mode:             graphql.Fm5SemanticModeAbsent,
-			EvidenceRevision: evidenceRevision,
-		}
-	} else {
-		verdict = deriveFM5Interpretation(
-			evidence.controller != 0,
-			moduleConfig,
-			solarReadable,
-			cylindersReadable,
-			evidence.hasEvidence() || currentEvidence.hasEvidence() || negativeIdentityObserved,
-			evidenceStale,
-			incoherent,
-			evidenceRevision,
-		)
-	}
-	verdict = p.commitFM5Acquisition(currentEvidence, verdict, incomingSolar, incomingCylinders)
-	for _, info := range fm5InventoryRegistryInfos(evidence.systemSnapshot, verdict.Mode) {
-		p.reg.Register(preserveExistingRegistryMetadata(p.reg, info))
-	}
-	p.publishFM5Semantic(semanticSnapshotSourceLive)
-}
-
-func (p *vaillantSemanticPoller) readSolarSnapshotStartup(ctx context.Context) (*vaillantSolarSnapshot, bool) {
-	incoming := &vaillantSolarSnapshot{}
-	readAny := false
-	if raw, ok := p.readB524Startup(ctx, localSolar.opcode, localSolar.group, solarInstance, solar_enabled); ok {
-		incoming.SolarEnabled = decodeB524BoolFromRaw(raw)
-		readAny = true
-	}
-	if raw, ok := p.readB524Startup(ctx, localSolar.opcode, localSolar.group, solarInstance, solar_collector_temp); ok {
-		incoming.CollectorTemperatureC = decodeB524Float32FromRaw(raw)
-		readAny = true
-	}
-	if raw, ok := p.readB524Startup(ctx, localSolar.opcode, localSolar.group, solarInstance, solar_pump_active); ok {
-		incoming.PumpActive = decodeB524BoolFromRaw(raw)
-		readAny = true
-	}
-	if !readAny {
-		return nil, false
-	}
-	return incoming, true
-}
-
-func (p *vaillantSemanticPoller) readCylinderSnapshotsStartup(ctx context.Context) (map[byte]*vaillantCylinderSnapshot, bool) {
-	out := make(map[byte]*vaillantCylinderSnapshot, 2)
-	for instance := byte(0x00); instance <= 0x01; instance++ {
-		raw, ok := p.readB524Startup(ctx, localCylinders.opcode, localCylinders.group, instance, cylinder_temperature)
-		if !ok {
-			continue
-		}
-		out[instance] = &vaillantCylinderSnapshot{
-			Instance:     instance,
-			TemperatureC: decodeB524Float32FromRaw(raw),
-		}
-	}
-	if len(out) == 0 {
-		return nil, false
-	}
-	return out, true
-}
-
-func (p *vaillantSemanticPoller) refreshBoilerStatusStartup(ctx context.Context) {
-	if p == nil {
-		return
-	}
-
-	p.mu.Lock()
-	boilerAddress := p.boilerAddress
-	p.mu.Unlock()
-	if boilerAddress == 0 {
-		boilerAddress = p.findBoilerAddress()
-	}
-
-	snapshot := &vaillantBoilerSnapshot{}
-	updated := false
-	if boilerAddress != 0 {
-		probeCtx, cancel := p.startupProbeContext(ctx)
-		if value := p.readB509DATA2c(probeCtx, boilerAddress, boiler_b509_flow_temperature); value != nil {
-			snapshot.FlowTemperatureC = value
-			updated = true
-		}
-		cancel()
-	}
-	if raw, ok := p.readB524Startup(ctx, localDHW.opcode, localDHW.group, dhwInstance, dhw_current_temp); ok {
-		if value := decodeB524Float32FromRaw(raw); value != nil {
-			snapshot.DhwTemperatureC = value
-			updated = true
-		}
-	}
-
-	p.mu.Lock()
-	if boilerAddress != 0 {
-		p.boilerAddress = boilerAddress
-	}
-	source := semanticSnapshotSourceCache
-	if updated {
-		p.boiler = mergeBoilerSnapshotNonDestructive(p.boiler, snapshot)
-		source = semanticSnapshotSourceLive
-	}
-	hasSnapshot := p.boiler != nil
-	p.mu.Unlock()
-	if hasSnapshot {
-		p.publishBoilerStatus(source)
-	}
-}
-
 func (p *vaillantSemanticPoller) Start(ctx context.Context) {
 	if p == nil {
 		return
@@ -2190,244 +430,6 @@ func (p *vaillantSemanticPoller) Start(ctx context.Context) {
 	}
 
 	p.startPollingLoops(ctx)
-}
-
-func (p *vaillantSemanticPoller) AttachPassiveShadowProducer(ctx context.Context, deduplicator *ebusgateway.ActivePassiveDeduplicator) error {
-	if p == nil || deduplicator == nil {
-		return nil
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-
-	subscription, err := deduplicator.Subscribe(
-		"semantic-vaillant-shadow",
-		passiveShadowSubscriberPriority,
-		passiveShadowSubscriberBuffer,
-	)
-	if err != nil {
-		return err
-	}
-
-	go func(subscription *ebusgateway.AdjudicatedPassiveSubscription) {
-		for {
-			closedUnexpectedly := false
-		consume:
-			for {
-				select {
-				case <-ctx.Done():
-					subscription.Close()
-					return
-				case event, ok := <-subscription.Events():
-					if !ok {
-						closedUnexpectedly = ctx.Err() == nil
-						break consume
-					}
-					p.handleAdjudicatedPassiveEvent(event)
-				}
-			}
-			subscription.Close()
-			if !closedUnexpectedly {
-				return
-			}
-			for {
-				if !waitForPassiveShadowRetry(ctx) {
-					return
-				}
-				next, err := deduplicator.Subscribe(
-					"semantic-vaillant-shadow",
-					passiveShadowSubscriberPriority,
-					passiveShadowSubscriberBuffer,
-				)
-				if err != nil {
-					continue
-				}
-				subscription = next
-				break
-			}
-		}
-	}(subscription)
-
-	return nil
-}
-
-func waitForPassiveShadowRetry(ctx context.Context) bool {
-	delay := passiveShadowRetryDelay
-	if delay <= 0 {
-		delay = 100 * time.Millisecond
-	}
-	timer := time.NewTimer(delay)
-	defer timer.Stop()
-
-	select {
-	case <-ctx.Done():
-		return false
-	case <-timer.C:
-		return true
-	}
-}
-
-func (p *vaillantSemanticPoller) handleAdjudicatedPassiveEvent(event ebusgateway.AdjudicatedPassiveEvent) {
-	if p == nil || p.shadow == nil {
-		return
-	}
-	if event.Disposition != ebusgateway.DedupDispositionUnmatchedThirdParty ||
-		event.ObservabilityOnly ||
-		event.LocalParticipantInbound ||
-		event.MatchedActiveDuplicate {
-		return
-	}
-
-	familyPolicy := event.Fingerprint.FamilyPolicy
-	if !passiveShadowLaneEnabled(p.shadow.FeatureFlags(), familyPolicy) {
-		return
-	}
-
-	key, ok := clonePassiveAdjudicatedWatchKey(event)
-	if !ok || key == nil {
-		return
-	}
-	runtime := p.bootstrapPassiveSharedWatchKey(key)
-
-	switch familyPolicy.RequestIntent {
-	case ebusgateway.ObserveFirstRequestIntentWrite:
-		if !shouldInvalidatePassiveExternalWrite(familyPolicy) {
-			return
-		}
-		invalidatedAt := event.Fingerprint.ObservedAt
-		if invalidatedAt.IsZero() {
-			invalidatedAt = p.now()
-		}
-		p.shadow.Invalidate(ebusgateway.ShadowInvalidation{
-			Key:           key,
-			Reason:        ebusgateway.ShadowInvalidationReasonExternalWrite,
-			Source:        ebusgateway.ShadowInvalidationSourcePassive,
-			InvalidatedAt: invalidatedAt,
-		})
-		return
-	case ebusgateway.ObserveFirstRequestIntentRead:
-	default:
-		return
-	}
-	if event.SuppressShadow {
-		return
-	}
-
-	if event.Fingerprint.ResponseClass != ebusgateway.DedupResponseValueBearing {
-		return
-	}
-	value, ok := parsePassiveShadowPayload(event, key)
-	if !ok || len(value) == 0 {
-		return
-	}
-
-	observedAt := event.Fingerprint.ObservedAt
-	if observedAt.IsZero() {
-		observedAt = p.now()
-	}
-	result := p.shadow.Write(ebusgateway.ShadowWrite{
-		Key:        key,
-		Source:     ebusgateway.ShadowWriteSourcePassive,
-		Confidence: ebusgateway.ShadowConfidenceHigh,
-		Value:      value,
-		ObservedAt: observedAt,
-	})
-	p.emitWatchDirectApplyEfficiency(runtime, observedAt, true, result.Accepted)
-	if !result.Accepted {
-		log.Printf("semantic_passive_shadow_write_rejected key=%q reason=%s", key.Canonical(), result.Reason)
-		return
-	}
-}
-
-func passiveShadowLaneEnabled(flags ebusgateway.ObserveFirstFeatureFlags, policy ebusgateway.ObserveFirstFamilyPolicy) bool {
-	if !flags.ObserveFirstEnabled() {
-		return false
-	}
-
-	switch policy.RequestIntent {
-	case ebusgateway.ObserveFirstRequestIntentRead:
-		switch policy.DirectApplyPolicy {
-		case ebusgateway.ObserveFirstDirectApplyPolicyStateDefault:
-			return flags.PassiveStateDirectApply()
-		case ebusgateway.ObserveFirstDirectApplyPolicyConfigOptIn:
-			return flags.PassiveConfigDirectApply()
-		default:
-			return false
-		}
-	case ebusgateway.ObserveFirstRequestIntentWrite:
-		if !policy.UsesRuntimeExternalWritePolicy {
-			return false
-		}
-		return policy.EffectiveExternalWritePolicy != ebusgateway.ObserveFirstExternalWritePolicyRecordOnly
-	default:
-		return false
-	}
-}
-
-func shouldInvalidatePassiveExternalWrite(policy ebusgateway.ObserveFirstFamilyPolicy) bool {
-	if !policy.UsesRuntimeExternalWritePolicy {
-		return false
-	}
-	switch policy.EffectiveExternalWritePolicy {
-	case ebusgateway.ObserveFirstExternalWritePolicyInvalidateOnly,
-		ebusgateway.ObserveFirstExternalWritePolicyRecordAndInvalidate:
-		return true
-	default:
-		return false
-	}
-}
-
-func clonePassiveAdjudicatedWatchKey(event ebusgateway.AdjudicatedPassiveEvent) (ebusgateway.WatchKey, bool) {
-	return cloneSemanticWatchKey(event.Fingerprint.SharedWatchKey)
-}
-
-func cloneSemanticWatchKey(key ebusgateway.WatchKey) (ebusgateway.WatchKey, bool) {
-	switch typed := key.(type) {
-	case ebusgateway.B509WatchKey:
-		cloned := ebusgateway.NewB509WatchKey(typed.Target, typed.RegisterAddress)
-		return cloned, true
-	case *ebusgateway.B509WatchKey:
-		if typed == nil {
-			return nil, false
-		}
-		cloned := ebusgateway.NewB509WatchKey(typed.Target, typed.RegisterAddress)
-		return cloned, true
-	case ebusgateway.B524WatchKey:
-		cloned := ebusgateway.NewB524WatchKey(typed.Target, typed.Opcode, typed.Group, typed.Instance, typed.RegisterAddress)
-		return cloned, true
-	case *ebusgateway.B524WatchKey:
-		if typed == nil {
-			return nil, false
-		}
-		cloned := ebusgateway.NewB524WatchKey(typed.Target, typed.Opcode, typed.Group, typed.Instance, typed.RegisterAddress)
-		return cloned, true
-	default:
-		return nil, false
-	}
-}
-
-func parsePassiveShadowPayload(event ebusgateway.AdjudicatedPassiveEvent, key ebusgateway.WatchKey) ([]byte, bool) {
-	if !event.Event.HasResponse {
-		return nil, false
-	}
-	switch typed := key.(type) {
-	case ebusgateway.B509WatchKey:
-		return parseB509ReadPayload(event.Event.Response.Data, typed.RegisterAddress)
-	case *ebusgateway.B509WatchKey:
-		if typed == nil {
-			return nil, false
-		}
-		return parseB509ReadPayload(event.Event.Response.Data, typed.RegisterAddress)
-	case ebusgateway.B524WatchKey:
-		return parseB524ReadPayload(event.Event.Response.Data, typed.Opcode, typed.Group, typed.Instance, typed.RegisterAddress)
-	case *ebusgateway.B524WatchKey:
-		if typed == nil {
-			return nil, false
-		}
-		return parseB524ReadPayload(event.Event.Response.Data, typed.Opcode, typed.Group, typed.Instance, typed.RegisterAddress)
-	default:
-		return nil, false
-	}
 }
 
 func (p *vaillantSemanticPoller) startPollingLoops(ctx context.Context) {
